Moved 'ManagedFeedImage', 'ManagedCache' and 'NSPersistentContainer' to its own classes
