+++ FeedStoreChallenge/CoreDataFeedStore.swift
@@ -66,85 +66,3 @@ public final class CoreDataFeedStore: FeedStore {
         }
     }
 }
-
-// MARK: Core Data Models
-
-@objc(ManagedCache)
-private class ManagedCache: NSManagedObject {
-    @NSManaged var timestamp: Date
-    @NSManaged var feed: NSOrderedSet
-    
-    var localFeed: [LocalFeedImage] {
-        return feed.compactMap { ($0 as? ManagedFeedImage)?.local }
-    }
-    
-    static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
-        let request = NSFetchRequest<ManagedCache>(entityName: ManagedCache.entity().name!)
-        request.returnsObjectsAsFaults = false
-        return try context.fetch(request).first
-    }
-    
-    static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
-        try find(in: context).map(context.delete)
-        return ManagedCache(context: context)
-    }
-}
-
-@objc(ManagedFeedImage)
-private class ManagedFeedImage: NSManagedObject {
-    @NSManaged var id: UUID
-    @NSManaged var imageDescription: String?
-    @NSManaged var location: String?
-    @NSManaged var url: URL
-    @NSManaged var cache: ManagedCache
-    
-    var local: LocalFeedImage {
-        return LocalFeedImage(id: id, description: imageDescription, location: location, url: url)
-    }
-
-    static func images(from localFeed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
-        return NSOrderedSet(array: localFeed.map { local in
-            let image = ManagedFeedImage(context: context)
-            image.id = local.id
-            image.imageDescription = local.description
-            image.location = local.location
-            image.url = local.url
-            return image
-        })
-    }
-}
-
-private extension NSPersistentContainer {
-
-    enum LoadError: Swift.Error {
-        case didNotFindModel
-        case didFailToLoadPersistentStores(Swift.Error)
-    }
-
-    static func load(modelName name: String, url: URL, in bundle: Bundle) throws -> NSPersistentContainer {
-        guard let model = NSManagedObjectModel.with(name: name, in: bundle) else {
-            throw LoadError.didNotFindModel
-        }
-
-        var loadError: Swift.Error?
-
-        let persistentDescription = NSPersistentStoreDescription(url: url)
-
-        let container = NSPersistentContainer(name: name, managedObjectModel: model)
-        container.persistentStoreDescriptions = [persistentDescription]
-
-        container.loadPersistentStores { (_, error) in
-            loadError = error
-        }
-
-        try loadError.map { throw LoadError.didFailToLoadPersistentStores($0) }
-
-        return container
-    }
-}
-
-private extension NSManagedObjectModel {
-    static func with(name: String, in bundle: Bundle) -> NSManagedObjectModel? {
-        return bundle.url(forResource: name, withExtension: "momd").flatMap { NSManagedObjectModel(contentsOf: $0) }
-    }
-}
