+++ FeedStoreChallenge/ManagedCache.swift
@@ -1,0 +1,31 @@
+//
+//  ManagedCache.swift
+//  FeedStoreChallenge
+//
+//  Created by Rakesh Ramamurthy on 22/10/20.
+//  Copyright © 2020 Essential Developer. All rights reserved.
+//
+
+import Foundation
+import CoreData
+
+@objc(ManagedCache)
+internal class ManagedCache: NSManagedObject {
+    @NSManaged internal var timestamp: Date
+    @NSManaged internal var feed: NSOrderedSet
+
+    internal var localFeed: [LocalFeedImage] {
+        return feed.compactMap { ($0 as? ManagedFeedImage)?.local }
+    }
+
+    internal static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
+        let request = NSFetchRequest<ManagedCache>(entityName: ManagedCache.entity().name!)
+        request.returnsObjectsAsFaults = false
+        return try context.fetch(request).first
+    }
+
+    internal static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
+        try find(in: context).map(context.delete)
+        return ManagedCache(context: context)
+    }
+}
